+++ app/src/main/java/com/example/newsgb/_core/ui/NavigationFragment.kt
@@ -24,6 +24,10 @@ class NavigationFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        initView()
+    }
+
+    private fun initView() {
         setFragmentInHostContainer(NewsFragment.newInstance())
         setBottomNavigationListener()
     }
