+++ app/src/main/java/com/example/newsgb/utils/ui/Categories.kt
@@ -1,9 +1,9 @@
-package com.example.newsgb.utils
+package com.example.newsgb.utils.ui
 
 import com.example.newsgb.R
 
-enum class Category(val categoryNameId: Int) {
-    GENERAL(R.string.general_category_name_text),
+enum class Category(val nameResId: Int) {
+    GENERAL(R.string.main_category_name_text),
     BUSINESS(R.string.business_category_name_text),
     SPORT(R.string.sports_category_name_text),
     ENTERTAINMENT(R.string.entertainment_category_name_text),
