+++ app/src/main/res/layout/news_fragment.xml
@@ -6,7 +6,6 @@
     android:layout_height="match_parent"
     tools:context="._core.ui.MainActivity">
 
-
     <com.google.android.material.tabs.TabLayout
         android:id="@+id/tab_layout"
         android:layout_width="match_parent"
@@ -17,14 +16,11 @@
         app:tabIndicatorColor="@color/category_button_text_color"
         app:tabMode="scrollable" />
 
-
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/view_pager"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="1"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tab_layout" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
