+++ app/src/main/res/layout/news_fragment_tab_item.xml
@@ -16,7 +16,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent">
 
-            <com.example.newsgb.utils.ui.EquilateralImageView
+            <com.example.newsgb.utils.widget.EquilateralImageView
                 android:id="@+id/first_news_image"
                 android:layout_width="match_parent"
                 android:layout_height="300dp"
